+++ src/main/java/uk/gov/hmcts/reform/pip/publication/services/service/artefactsummary/ArtefactSummaryService.java
@@ -1,10 +1,11 @@
-package uk.gov.hmcts.reform.pip.publication.services.service;
+package uk.gov.hmcts.reform.pip.publication.services.service.artefactsummary;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.pip.publication.services.models.external.ListType;
 
@@ -25,13 +26,14 @@ import java.util.Locale;
  */
 public class ArtefactSummaryService {
 
+    @Autowired
+    SjpPublicList sjpPublicList;
+
     private static final String COURT_LISTS = "courtLists";
     private static final String COURT_HOUSE = "courtHouse";
     private static final String COURT_ROOM = "courtRoom";
     private static final String SESSION = "session";
-    private static final String OFFENCE = "offence";
     private static final String SITTINGS = "sittings";
-    private static final String HEARING = "hearing";
     private static final String INDIVIDUAL_DETAILS = "individualDetails";
 
     /**
@@ -46,7 +48,7 @@ public class ArtefactSummaryService {
     public String artefactSummary(String payload, ListType listType) throws JsonProcessingException {
         switch (listType) {
             case SJP_PUBLIC_LIST:
-                return artefactSummarysjpPublic(payload);
+                return sjpPublicList.artefactSummarysjpPublic(payload);
             case SJP_PRESS_LIST:
                 return artefactSummarysjpPress(payload);
             case CIVIL_DAILY_CAUSE_LIST:
@@ -267,9 +269,8 @@ public class ArtefactSummaryService {
         boolean restriction = node.get("reportingRestriction").asBoolean();
         if (restriction) {
             return "(Reporting restriction)";
-        } else {
-            return "";
         }
+        return "";
     }
 
     /**
@@ -297,56 +298,5 @@ public class ArtefactSummaryService {
         return "Accused: " + accused + "\nPostcode: " + postCode + "\nProsecutor: " + prosecutor;
     }
 
-    /**
-     * parent method for sjp public lists. iterates on sittings.
-     *
-     * @param payload - json body.
-     * @return string of data.
-     * @throws JsonProcessingException - jackson prereq.
-     */
-    private String artefactSummarysjpPublic(String payload) throws JsonProcessingException {
-        StringBuilder output = new StringBuilder();
-        JsonNode node = new ObjectMapper().readTree(payload);
-        Iterator<JsonNode> sittings =
-            node.get(COURT_LISTS).get(0).get(COURT_HOUSE).get(COURT_ROOM).get(0)
-                .get(SESSION).get(0).get(SITTINGS).elements();
-        while (sittings.hasNext()) {
-            output.append('•');
-            JsonNode currentHearing = sittings.next();
-            output.append(processRolesSjpPublic(currentHearing));
-            String offence = currentHearing.get(HEARING).get(0).get(OFFENCE).get(0).get("offenceTitle").asText();
-            output.append("Offence: ").append(offence).append('\n');
-        }
-        return output.toString();
-    }
 
-    /**
-     * handle sjp public roles iteration.
-     * @param hearing - node of a given hearing.
-     * @return string of roles.
-     */
-    private String processRolesSjpPublic(JsonNode hearing) {
-        StringBuilder outputString = new StringBuilder();
-        Iterator<JsonNode> partyNames = hearing.get(HEARING).get(0).get("party").elements();
-        while (partyNames.hasNext()) {
-            JsonNode currentParty = partyNames.next();
-            switch (currentParty.get("partyRole").asText()) {
-                case "ACCUSED":
-                    String forenames = currentParty.get(INDIVIDUAL_DETAILS).get("individualForenames").asText();
-                    String surname = currentParty.get(INDIVIDUAL_DETAILS).get("individualSurname").asText();
-                    String postCode = currentParty.get(INDIVIDUAL_DETAILS).get("address").get("postCode").asText();
-                    outputString.append("Accused: ").append(forenames).append(' ').append(surname);
-                    outputString.append("\nPostcode: ").append(postCode).append('\n');
-                    break;
-                case "PROSECUTOR":
-                    outputString.append("Prosecutor: ")
-                        .append(currentParty.get("organisationDetails").get("organisationName").asText())
-                        .append('\n');
-                    break;
-                default:
-                    break;
-            }
-        }
-        return outputString.toString();
-    }
 }
